refactor(tangle-dapp): Move useIsMountedRef to webb-ui-components
