+++ apps/tangle-dapp/hooks/useSubstrateTx.ts
@@ -5,6 +5,7 @@ import { PromiseOrT } from '@webb-tools/abstract-api-provider';
 import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
 import ensureError from '@webb-tools/tangle-shared-ui/utils/ensureError';
 import { getApiPromise } from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
+import useIsMountedRef from '@webb-tools/webb-ui-components/hooks/useIsMountedRef';
 import assert from 'assert';
 import { useCallback, useEffect, useState } from 'react';
 import { Hash } from 'viem';
@@ -14,7 +15,6 @@ import { GetSuccessMessageFunction } from '../types';
 import extractErrorFromTxStatus from '../utils/extractErrorFromStatus';
 import useActiveAccountAddress from './useActiveAccountAddress';
 import useAgnosticAccountInfo from './useAgnosticAccountInfo';
-import useIsMountedRef from './useIsMountedRef';
 import useSubstrateAddress from './useSubstrateAddress';
 import useSubstrateExplorerUrl from './useSubstrateExplorerUrl';
 import useSubstrateInjectedExtension from './useSubstrateInjectedExtension';
